fix: torch_dtype= → dtype= in online_serve.py model loading
transformers >= 4.51 deprecates torch_dtype= in from_pretrained();
dtype= is the correct kwarg for Qwen3. Both target and draft base
model loads updated. This was causing a deprecation warning on every
online_ebe_adapt run.

+++ gbv-research/algorithms/online_serve.py
@@ -655,7 +655,7 @@ def load_models(args: argparse.Namespace, device: torch.device):
     else:
         target_model = AutoModelForCausalLM.from_pretrained(
             args.target,
-            torch_dtype=dtype,
+            dtype=dtype,          # transformers ≥ 4.51 (torch_dtype= deprecated)
             device_map=device,
         )
     target_model.eval()
@@ -665,7 +665,7 @@ def load_models(args: argparse.Namespace, device: torch.device):
     log.info("Loading draft base model: %s", args.draft)
     draft_base = AutoModelForCausalLM.from_pretrained(
         args.draft,
-        torch_dtype=dtype,
+        dtype=dtype,              # transformers ≥ 4.51 (torch_dtype= deprecated)
     ).to(device)
 
     # Auto-resume: if a ckpt_latest/ exists in the output dir use it as the
